feat: 🎸 Switch to using cornerstone prefetchers
Switched to using cornerstone prefetchers so we don't send 1000 requests
at the same time. This is a breaking change to the API as we switch to a
sub model for the events.

BREAKING CHANGE: The imageDataObject no-longer has insertPixelDataPromises, insteads the
consumer may subscribe to an imageDataObject's onPixelDataInserted
and/or onAllPixelDataInserted.